feat: 重新设计JwtClaims结构体和CommonUser结构体，删除原先的Claims结构体

+++ infra/src/conf/configuration.rs
@@ -4,7 +4,7 @@ use sqlx::postgres::{PgConnectOptions, PgSslMode};
 use std::collections::HashMap;
 use std::path::PathBuf;
 
-pub type TokenConfig = HashMap<String, i16>;
+pub type TokenConfig = HashMap<String, i64>;
 
 ///---------------------- 顶层 Setting ----------------------
 #[derive(serde::Deserialize, Clone, Debug)]
